+++ code/src/ui/graphic/table/registry/ribbon/table-ribbon.tsx
@@ -12,7 +12,6 @@ import RedirectButton from "ui/interaction/action/redirect/redirect-button";
 import ReturnButton from "ui/interaction/action/redirect/return-button";
 import Button from "ui/interaction/button";
 import DateRangeInput from "ui/interaction/input/date-range";
-import ColumnSearchComponent from "../actions/column-search";
 import { DateRange } from "react-day-picker";
 
 interface TableRibbonProps {
@@ -55,37 +54,46 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
         (!keycloakEnabled ||
           !permissionScheme ||
           permissionScheme.hasPermissions.allTasks) && (
-          <div className="flex flex-col sm:flex-row items-center justify-between sm:gap-4 bg-gray-200 dark:bg-zinc-800 w-full sm:max-w-fit p-1.5 text-center rounded-lg gap-2">
-            <div className="w-full sm:w-auto border-b-1 border-gray-400 pb-2 sm:p-0 sm:border-0">
-              <RedirectButton
-                label={dict.nav.title.outstanding}
-                leftIcon="pending"
-                url={`${Routes.REGISTRY_TASK_OUTSTANDING}`}
-                variant={
-                  props.lifecycleStage == "outstanding" ? "active" : "ghost"
-                }
-                className="w-full"
-              />
-            </div>
-            <div className="w-full sm:w-auto border-b-1 border-gray-400 pb-2 sm:p-0 sm:border-0">
-              <RedirectButton
-                label={dict.nav.title.scheduled}
-                leftIcon="schedule"
-                url={`${Routes.REGISTRY_TASK_SCHEDULED}`}
-                variant={
-                  props.lifecycleStage == "scheduled" ? "active" : "ghost"
-                }
-                className="w-full"
-              />
-            </div>
-            <div className="w-full sm:w-auto">
-              <RedirectButton
-                label={dict.nav.title.closed}
-                leftIcon="archive"
-                url={`${Routes.REGISTRY_TASK_CLOSED}`}
-                variant={props.lifecycleStage == "closed" ? "active" : "ghost"}
-                className="w-full"
-              />
+          <div className="bg-ring w-full sm:max-w-fit rounded-lg p-2 sm:p-1.5">
+            <div className="flex  sm:items-center sm:justify-between sm:gap-4 gap-1">
+              <div className="w-full sm:w-auto">
+                <RedirectButton
+                  label={dict.nav.title.outstanding}
+                  leftIcon="pending"
+                  hasMobileIcon={false}
+                  url={`${Routes.REGISTRY_TASK_OUTSTANDING}`}
+                  variant={
+                    props.lifecycleStage == "outstanding" ? "active" : "ghost"
+                  }
+                  className="w-full sm:w-auto py-3 sm:py-2 text-sm font-medium"
+                />
+              </div>
+
+              <div className="w-full sm:w-auto">
+                <RedirectButton
+                  label={dict.nav.title.scheduled}
+                  leftIcon="schedule"
+                  hasMobileIcon={false}
+                  url={`${Routes.REGISTRY_TASK_SCHEDULED}`}
+                  variant={
+                    props.lifecycleStage == "scheduled" ? "active" : "ghost"
+                  }
+                  className="w-full sm:w-auto py-3 sm:py-2 text-sm font-medium"
+                />
+              </div>
+
+              <div className="w-full sm:w-auto">
+                <RedirectButton
+                  label={dict.nav.title.closed}
+                  leftIcon="archive"
+                  hasMobileIcon={false}
+                  url={`${Routes.REGISTRY_TASK_CLOSED}`}
+                  variant={
+                    props.lifecycleStage == "closed" ? "active" : "ghost"
+                  }
+                  className="w-full sm:w-auto py-3 sm:py-2 text-sm font-medium"
+                />
+              </div>
             </div>
           </div>
         )}
@@ -99,12 +107,6 @@ export default function TableRibbon(props: Readonly<TableRibbonProps>) {
             variant="outline"
             onClick={triggerRefresh}
           />
-          {props.instances.length > 1 && (
-            <ColumnSearchComponent
-              instances={props.instances}
-              setCurrentInstances={props.setCurrentInstances}
-            />
-          )}
         </div>
         <div className="flex  flex-wrap gap-2 mt-2 md:mt-0  ">
           {(!keycloakEnabled ||