+++ code/src/ui/interaction/header/headerbar.tsx
@@ -67,10 +67,7 @@ export default function HeaderBar(props: Readonly<HeaderBarProps>) {
           {Array.isArray(props.settings?.branding?.navbar) ? (
             props.settings?.branding?.navbar.map((logo) => (
               <Link key={logo} href={Routes.HOME}>
-                <IconComponent
-                  icon={logo}
-                  classes="h-8 md:h-10 lg:h-10 2xl:h-12 w-auto ml-2 md:ml-8"
-                />
+                <IconComponent icon={logo} classes="h-12 w-auto ml-2 md:ml-8" />
               </Link>
             ))
           ) : (
@@ -78,7 +75,7 @@ export default function HeaderBar(props: Readonly<HeaderBarProps>) {
             <Link href={Routes.HOME}>
               <IconComponent
                 icon={props.settings?.branding?.navbar}
-                classes="h-8 md:h-10 lg:h-10 2xl:h-12 ml-4 w-auto"
+                classes="h-12 ml-4 w-auto"
               />
             </Link>
           )}